[REFACTOR] 마이페이지관련 모델분리(#173)
