+++ LionHeart-iOS/LionHeart-iOS/Scenes/MyPage/Model/MyPageSection.swift
@@ -1,0 +1,13 @@
+//
+//  MyPageSection.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/11/20.
+//
+
+import Foundation
+
+enum MyPageSection: String, CaseIterable {
+    case customerServiceSetion = "고객센터"
+    case appSettingSection = "앱 설정"
+}
