+++ LionHeart-iOS/LionHeart-iOS/Scenes/MyPage/Model/MyPageSectionModel.swift
@@ -1,0 +1,12 @@
+//
+//  MyPageSectionModel.swift
+//  LionHeart-iOS
+//
+//  Created by uiskim on 2023/11/20.
+//
+
+import Foundation
+
+struct MyPageSectionModel: Hashable {
+    let cellTitle: String
+}
